+++ lib/bothan/helpers/auth_helpers.rb
@@ -1,0 +1,18 @@
+module Bothan
+  module Helpers
+    module Auth
+
+      def protected!
+        return if authorized?
+        headers['WWW-Authenticate'] = 'Basic realm="Restricted Area"'
+        halt 401, "Not authorized\n"
+      end
+
+      def authorized?
+        @auth ||=  Rack::Auth::Basic::Request.new(request.env)
+        @auth.provided? and @auth.basic? and @auth.credentials and @auth.credentials == [ENV['METRICS_API_USERNAME'], ENV['METRICS_API_PASSWORD']]
+      end
+
+    end
+  end
+end
