+++ spec/extensions/string_spec.rb
